Initial_code/requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.27
Pillow==10.2.0
python-dotenv==1.0.1 